backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]
pydantic
python-dotenv
google-generativeai
